change pipette to instrument and container to placeable, fix anchor tag href
