--- app/renderer/src/components/Pipette.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>
-    Calibrate {{ $route.params.pipette }}
-  </div>
-</template>
-
-<script>
-  export default {
-    name: "Pipette"
-  }
-</script>
