+++ app/renderer/src/components/export.js
@@ -2,5 +2,5 @@ export {default as Home} from './Home.vue'
 export {default as Connect} from './Connect.vue'
 export {default as StepList} from './StepList.vue'
 export {default as Upload} from './Upload.vue'
-export {default as Container} from './Container.vue'
-export {default as Pipette} from './Pipette.vue'
+export {default as Placeable} from './Placeable.vue'
+export {default as Instrument} from './Instrument.vue'
